Devel

+++ src/mkplst.f
@@ -14,7 +14,7 @@ c
 c      allmem=mtaloc(-1)
       mtop=ktaloc(allmem)
       mlen=0
-      call push(0,0,mlen,ilist(1,mtop),allmem)
+      call push(int8(0),0,mlen,ilist(1,mtop),allmem)
 c
       level=0
  1000 continue
@@ -28,7 +28,7 @@ c
         call pop(i,llen,isp,ilist(1,pstack),STKSIZ)
       else
         if(lstchk(ilist(2,i),ilist(1,mtop),mlen,allmem) .eq. 0) then
-          call push(0,ilist(2,i),mlen,ilist(1,mtop),allmem)
+          call push(int8(0),ilist(2,i),mlen,ilist(1,mtop),allmem)
         endif
         if (idtype(ilist(2,i)) .eq. icLINE) then
           call push(i,llen,isp,ilist(1,pstack),STKSIZ)
